Refactored DTOs and overall cleanup of the code.
