+++ API/Controllers/AccountController.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using System.Threading.Tasks;
 using API.Constants;
 using API.DTOs;
+using API.DTOs.Account;
 using API.Entities;
 using API.Errors;
 using API.Extensions;
