+++ API/Controllers/ReaderController.cs
@@ -483,6 +483,8 @@ namespace API.Controllers
                     && x.Page != bookmarkDto.Page).ToList();
 
 
+
+
                 _unitOfWork.UserRepository.Update(user);
 
                 if (await _unitOfWork.CommitAsync())
