+++ API/DTOs/Account/ResetPasswordDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace API.DTOs
+namespace API.DTOs.Account
 {
     public class ResetPasswordDto
     {
