+++ API/DTOs/Reader/RemoveBookmarkForSeriesDto.cs
@@ -1,4 +1,4 @@
-﻿namespace API.DTOs
+﻿namespace API.DTOs.Reader
 {
     public class RemoveBookmarkForSeriesDto
     {
