+++ UI/Web/src/app/_models/series.ts
@@ -8,7 +8,6 @@ export interface Series {
     localizedName: string;
     sortName: string;
     summary: string;
-    coverImage: string; // This is not passed from backend any longer. TODO: Remove this field
     coverImageLocked: boolean;
     volumes: Volume[];
     pages: number; // Total pages in series
